+++ lib/core/validation/validators/validator-required-when-any.js
@@ -3,13 +3,13 @@
 var _ = require('lodash');
 var $ = require('jquery');
 var moment = require('moment');
-require('./validator.required.when');
+require('./validator-required-when');
 
 (function(root) {
 
   var availity = root.availity;
 
-  availity.core.factory('avValRequiredWhenAll', function(avValUtils, avValRequiredWhen) {
+  availity.core.factory('avValRequiredWhenAny', function(avValUtils, avValRequiredWhen) {
 
     function createRequiredWhenRule(index, field) {
       var rule = {};
@@ -42,7 +42,7 @@ require('./validator.required.when');
     }
 
     var validator =  {
-      name: 'requiredWhenAll',
+      name: 'requiredWhenAny',
       validate: function(value, rules) {
         var valid = true;
         var replacementRegex = /^\$[a-zA-Z_0-9.]*/;
@@ -67,15 +67,15 @@ require('./validator.required.when');
                       }
                       if (textField) {
                         testValue = calculateOffsets(textField, offsetValue[0]);
-                        getField[value] = testValue;
+                        getField[fieldValue] = testValue;
                       }
                     }
                   }
                 }
                 // Invoke the requiredWhen - the subsitutions have been made.
-                if (valid) {
+                if (!valid) {
                   var requiredWhenRule = createRequiredWhenRule(index, getField);
-                  valid = avValRequiredWhen.validate(value, requiredWhenRule);
+                  valid = avValRequiredWhen.validate(index, requiredWhenRule);
                 }
               }
             }
