+++ lib/core/validation/validators/validator-required-when.js
@@ -2,11 +2,11 @@
 
 var _ = require('lodash');
 var $ = require('jquery');
-require('./validator.equal.to');
-require('./validator.less.equal');
-require('./validator.less.than');
-require('./validator.greater.equal');
-require('./validator.greater.than');
+require('./validator-equal-to.js');
+require('./validator-less-equal.js');
+require('./validator-less-than.js');
+require('./validator-greater-equal.js');
+require('./validator-greater-than.js');
 
 (function(root) {
